requirements.txt
pandas
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.1
Werkzeug==0.14.1
azure-storage-blob
azure-keyvault-secrets
azure-identity
dash
dash-renderer
dash-html-components
dash-core-components
plotly
mysql-connector-python
sqlalchemy